+++ lib/captcha.dart
@@ -1,19 +1,20 @@
-library capatcha;
+library captcha;
 
-import 'dart-ext:capatcha';
+import 'dart-ext:captcha';
 import 'dart:math';
 import 'dart:io';
 import 'dart:async';
+import 'dart:collection';
 
 typedef String TextGenerator();
 
-typedef void CapatchaGenerated(Capatcha capatcha);
+typedef void CaptchaGenerated(Captcha captcha);
 
 void _generate(String captchaText, String fileOut, int count, int width, int height, int offset, int quality, int isjpeg, int fontSize) native 'Generate';
 
-class Capatcha {
+class Captcha {
 
-  _Capatcha() {
+  _Captcha() {
   }
 
   String _text;
@@ -28,11 +29,11 @@ class Capatcha {
     return _file;
   }
 
-  static final Arrays<String> charDic = ['0', '1', '2', '3', '4', '5', '6', '7', '8', '9', 'A', 'B', 'C', 'D', 'E', 'F', 'G', 'H', 'I', 'J', 'K', 'L', 'M', 'N', 'O', 'P', 'Q', 'R', 'S', 'T', 'U', 'V', 'W', 'X', 'Y', 'Z'];
+  static final List<String> charDic = ['0', '1', '2', '3', '4', '5', '6', '7', '8', '9', 'A', 'B', 'C', 'D', 'E', 'F', 'G', 'H', 'I', 'J', 'K', 'L', 'M', 'N', 'O', 'P', 'Q', 'R', 'S', 'T', 'U', 'V', 'W', 'X', 'Y', 'Z'];
 
   static final Random random = new Random();
 
-  static Future<Capatcha> generate({TextGenerator textGenerator, CapatchaGenerated generated, String fileOut, int width:230, int height:60, int offset:35, int quality:50, int fontSize:57, bool removeFile:true}) {
+  static Future<Captcha> generate({TextGenerator textGenerator, CaptchaGenerated generated, String fileOut, int width:230, int height:60, int offset:35, int quality:50, int fontSize:57, bool removeFile:true}) {
     var completer = new Completer();
     var str_num = 6, r_num = charDic.length, text = '';
     if (textGenerator == null) {
@@ -54,7 +55,7 @@ class Capatcha {
     }
 
     _generate(text, fileOut, text.length, width, height, offset, quality, fileOut.endsWith('.jpg') || fileOut.endsWith('.jpeg') ? 1 : 0, fontSize);
-    var c = new Capatcha();
+    var c = new Captcha();
     c._text = text;
     c._file = new File(fileOut);
     try {